+++ src/Controller/MessageRestController.php
@@ -6,6 +6,7 @@ use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Message\MessageModel;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\FOSRestController;
+use FOS\RestBundle\Request\ParamFetcher;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class MessageRestController extends FOSRestController
@@ -43,13 +44,18 @@ class MessageRestController extends FOSRestController
 	}
 
 	/**
-	 * @Rest\Get("conversations/{conversationId}/{messagesNumber}/{messagesOffset}", requirements={"conversationId" = "\d+"}, defaults={"messagesNumber" = 20, "messagesOffset" = 0})
+	 * @Rest\Get("conversations/{conversationId}", requirements={"conversationId" = "\d+"}, defaults={"messagesNumber" = 20})
+	 * @Rest\QueryParam(name="messagesLimit", requirements="\d+", default="20", description="How many messages to return.")
+	 * @Rest\QueryParam(name="messagesOffset", requirements="\d+", default="0", description="Offset returned messages.")
 	 */
-	public function getConversationAction(int $conversationId, int $messagesNumber, int $messagesOffset)
+	public function getConversationAction(int $conversationId, ParamFetcher $paramFetcher)
 	{
 		if (!$this->session->may() || !$this->mayConversation($conversationId)) {
 			throw new HttpException(401);
 		}
+
+		$messagesLimit = $paramFetcher->get('messagesLimit');
+		$messagesOffset = $paramFetcher->get('messagesOffset');
 
 		$member = $this->model->listConversationMembers($conversationId);
 		$publicMemberInfo = function ($member) {
@@ -61,7 +67,7 @@ class MessageRestController extends FOSRestController
 		};
 		$member = array_map($publicMemberInfo, $member);
 
-		$messages = $this->model->loadConversationMessages($conversationId, $messagesNumber, $messagesOffset);
+		$messages = $this->model->loadConversationMessages($conversationId, $messagesLimit, $messagesOffset);
 		$conversation = $this->model->getValues(array('name'), 'conversation', $conversationId);
 		$this->model->setAsRead([$conversationId]);
 
@@ -72,6 +78,26 @@ class MessageRestController extends FOSRestController
 		];
 
 		$view = $this->view($data, 200);
+
+		return $this->handleView($view);
+	}
+
+	/**
+	 * @Rest\Get("conversations")
+	 * @Rest\QueryParam(name="limit", requirements="\d+", default="20", description="How many conversations to return.")
+	 * @Rest\QueryParam(name="offset", requirements="\d+", default="0", description="Offset returned conversations.")
+	 */
+	public function getConversationsAction(ParamFetcher $paramFetcher)
+	{
+		if (!$this->session->may()) {
+			throw new HttpException(401);
+		}
+
+		$limit = $paramFetcher->get('limit');
+		$offset = $paramFetcher->get('offset');
+
+		$conversations = $this->model->listConversations($limit, $offset);
+		$view = $this->view($conversations, 200);
 
 		return $this->handleView($view);
 	}